BUG FIXES * Now functioning: in conclusion of Manage Booking process - payment via C-O-D, delete cookies-on-server. * In EventCtrl/confirm_step2, when creating B C-O-S, we forgot to set the ticket class unique ID so therefore when finally confirming ( accessing EventCtrl/confirm_step3 via AJAX ), can't confirm. * In EventCtrl/confirm_step3, the feedback is not sent to client when confirmation is not successfully processed. * MB view and cancel pending changes now working. * Formerly: If attempted PayPal payment, then decided to cancel - will be diverted always to EventCtrl/book_step5_forward instead of EventCtrl/managebooking_confirm when appropriate. ( Noteworthy bug if this is not fixed: If we allow that, and then the user cancelled the process, the booking is now settled - marked as paid, the old slots and seats removed and the new ones now assigned to that booking! ). * In Manage Booking Confirmation/Review page: cancelling is now enabled - even if we are just changing payment modes. * Manage Booking - Change Payment mode now implemented. * Confirmation page on whether the user would like to choose seats for the guests or not is now working. And this is not shown if the new target showtime requires that a seat be chosen when booking. * When doing changes to a booking and some seats of guests cannot be automatically assigned to the new booking configuration, the appropriate purchase items might not be created formerly. * In previous releases, When MB > new st > same class > some guests' seats not available > seat selection page - the seat of the guest shouldn't have been reserved but is displayed as reserved. * Formerly, when user went to seat selection page while in same ticket class (i.e. clicked Change seat in MB home page) - and then did not change any , will get a "Seat occuppied" warning. Fix, in seatmaintenance/areSeatsOccuppied: If submitted seat is the same as the old seat, skipped for availability check as before coming to seat selection page, the seats are temporarily assigned to the guests beforehand. * JS Warning if a showing time is not selected in EventCtrl/managebooking_changeshowingtime. * Formerly, if seats have already been chosen > MB > Change seat > remove assigned/chosen seat > seat is not removed from user. * Formerly, when just changing seats in Manage Booking and then cancelling the process - the seats of the entire booking are freed ( but the corresponding columns  in `event_slot` is not touched ). * Cancelling in Manage Booking in the showing time selection page when changing showing time is now also supported. * Finally, fixed problem with assets/images/footer-bg.PNG uppercase hell. :D
CHANGES IN FEATURES/ALGORITHMS
* The use of MYSQL's CURRENT_TIMESTAMP is deprecated in favor of PHP's date()
so as to minimize errors/inconsistencies regarding time zone differences in web hosting.
* In paypal payments, when sending mail sales email is now fetched from DB, and sender name is got
from the email model via the respective constant.
* Account_model/isUserAuthorizedPaymentAgency does not check anymore if a payment mode is ONLINE - 
that task should be delegated to the calling controller/library now.
* booking_model::isBookingUnderThisUser returns NULL if the booking number to be checked is not found in the DB.
* When making a new booking (that is, not MB), the booking's `Status2` field is updated.
* When accessing Manage Booking on a new tab while you are booking a new reservation in another,
you will be redirected to the page where you are doing the booking.
* (views) html-generic/customer_seat_info.inc now utilized in confirmReservation/confirmReservation02-free.php, book_step6_cod.php 
confirmReservation/confirmReservation02.php
* (views) html-generic/customer_proper_info.inc" now utilized in confirmReservation/confirmReservation02-free.php, bookStep4.php, book_step6_cod.php 
confirmReservation/confirmReservation02.php
* assets\javascript\seatV2\seatManipulation.js::assembleExistingGuestSeatData : hidden input field whose name is g<guestNum>_uuid is now not considered,
 and the corresponding array index for the guest who does not have a seat assignment is set to boolean FALSE.
* assets\javascript\seatV2\seatManipulation.js::makeSeatUsedByThisBookingAvailable, the matrixRep is checked first if boolean FALSE before processing.
* library seatmaintenance::getExistingSeatData_ForManageBooking deprecated and removed in favor of:
* EvenCtrl/getSeatRepresentationsOfGuests moved to library seatmaintenance
* Entering just a single character in EventCtrl/book_step3::get-info-from-profile will now give a modal warning.
* When making changes to booking, the field `seats_actual`.`Comments` are updated with the expiration date of the new booking C-O-S.
When the changes are finalized - they are again updated with the deadline for the payment if any, for the changes in the booking.

STILL UP NEXT
* VERY HIGH | Column markers in seat selection is not properly working/stated for the real D.L. Umali Hall seat map|
Overhaul seat architecture ( see uxt-onwork-debug.txt )
* MEDIUM | Lowercase them all! ( all controllers' and all models' file names )
* MEDIUM | application/views/createClass/createClass02.php - why does it need 'assets/javascript/form-validation/manageBooking02.js?
* MEDIUM | Unify email module
* MEDIUM | Email new booking too if free
* MEDIUM| IN mb, choose seat "Your current ticket class is " - ang nadidisplay eh yung new ticket class instead of the old.
* LOW | In book_step3_forward - when getting user info from profile - display some ajax indicator and disable all inputs. Upon end of ajax ca;;,
re-enable.
* LOW | In MakeXML_model - disable file writing on functions not needed.
* LOW | Implement EventCtrl/manageBooking - where you see the archived bookings!
* LOW| HTML mail!
* LOW| Make "View Details" work for a new booking pending payment.
* LOW| Resume booking functionality.
* LOW| Encode in functionaccess::getRedirectURL - the stages and their corresponding URIs.
* VERY LOW | Decide whether to update the `Status` and `Status2` field when managing a booking.
* VERY LOW| In confirm_step3() (as well as in book_step6 and managebooking_finalize):
update transactionlist's data and datatype after creating payment - set the
payment uniqueID there.
* VERY LOW | When paypal IPN received but does not pass our standards (i.e.
$this->Payment_model->isPaypalPaymentOK() returns FALSE, notify user and act.
* VERY LOW| When paypal IPN received and bookingmaintenance/pay_and_confirm returns FALSE,
email user about the error/notify them when returning to the site.
* VERY LOW| Deny listing of contents of assets/xmltemp if deemed no other functionality is using that.

+++ trunk/application/models/ndx_model.php
@@ -79,9 +79,20 @@ class ndx_model extends CI_Model {
 		*	@created 09JUN2012-1244
 		*	@description Well, obviously, deletes.
 		*	@returns BOOLEAN - whether transaction was carried out successfully (TRUE) or not (FALSE)
-		**/		
+		**/
+		if( is_null( $uuid ) or $uuid === FALSE ) return TRUE;
 		return $this->db->delete( COL_DB_TABLE_NAME, Array( COL_UUID => $uuid ) );
 	}//delete(..)
+	
+	function deleteExpiredBookingCookiesOnServer(){
+		/**
+		*	@created 24JUN2012-1916
+		*	@description Deletes any booking cookies-on-server that is expired.
+		**/
+		$sql_command = "DELETE FROM `" . COL_DB_TABLE_NAME . "` WHERE CONCAT(`";
+		$sql_command .= COL_EXPIRE_DATE . "`,' ',`".COL_EXPIRE_TIME . "`) <= ? ";
+		return $this->db->query( $sql_command, Array( date("Y-m-d H:i:s"),  ) );
+	}//deleteExpiredBookingCookiesOnServer()
 	
 	function get( $uuid )
 	{
